Refactor weekend day card into component
